fix: bug ById corrigido

+++ backend_elgeladon/src/controllers/paleta.controller.js
@@ -1,39 +1,61 @@
 const paletasService = require('../services/paleta.service');
 
 const findAllPaletasController = (req, res) => {
-    const paletas = paletasService.findAllPaletasService();
-    res.send(paletas);
+  const paletas = paletasService.findAllPaletasService();
+  res.send(paletas);
 };
 
 const findByIdPaletasController = (req, res) => {
-    const idParam = +req.params.id;
+  const idParam = +req.params.id;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
   const chosenPaleta = paletasService.findByIdPaletasService(idParam);
   res.send(chosenPaleta);
 };
 
 const createPaletaController = (req, res) => {
-    const paleta = req.body;
-    const newPaleta = paletasService.createPaletaService(paleta);
-    res.send(newPaleta);
-}
+  const paleta = req.body;
+  if (
+    !paleta ||
+    !paleta.sabor ||
+    !paleta.descricao ||
+    !paleta.foto ||
+    !paleta.preco
+  ) {
+    return res.send({ mensagem: "Você não preencheu todos os dados para adicionar uma nova paleta ao cardápio!" });
+  }
+  const newPaleta = paletasService.createPaletaService(paleta);
+  res.send(newPaleta);
+};
 
 const updatePaletaController = (req, res) => {
-    const idParam = +req.params.id;
-    const paletaEdit = req.body;
-    const updatedPaleta = paletasService.updatePaletaService(idParam, paletaEdit);
-    res.send(updatedPaleta);
-}
+  const idParam = +req.params.id;
+  const paletaEdit = req.body;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
+
+  if (!paletaEdit || !paletaEdit.sabor || !paletaEdit.descricao || !paletaEdit.foto || !paletaEdit.preco) {
+    return res.status(400).send({ message: "Você não preencheu todos os dados para editar a paleta!" });
+  }
+  const updatedPaleta = paletasService.updatePaletaService(idParam, paletaEdit);
+  res.send(updatedPaleta);
+};
 
 const deletePaletaController = (req, res) => {
-    const idParam = req.params.id;
-    paletasService.deletePaletaService(idParam);
-    res.send({ message: 'Paleta deletada com sucesso!' });
-}
+  const idParam = req.params.id;
+  if (!idParam) {
+    return res.status(404).send({ message: "Paleta não encontrada!" })
+  }
+  paletasService.deletePaletaService(idParam);
+  res.send({ message: 'Paleta deletada com sucesso!' });
+};
 
 module.exports = {
-    findAllPaletasController,
-    findByIdPaletasController,
-    createPaletaController,
-    updatePaletaController,
-    deletePaletaController
-}
+  findAllPaletasController,
+  findByIdPaletasController,
+  createPaletaController,
+  updatePaletaController,
+  deletePaletaController,
+};
